moved the menu to a separate class
