+++ app/src/main/java/com/edurda77/filmlibrary/data/LocationRepoImpl.kt
@@ -13,16 +13,16 @@ import com.google.android.gms.maps.model.MarkerOptions
 class LocationRepoImpl: LocationRepo {
 
 
-    override fun setLocation(place: String, context: Context, mMap: GoogleMap) {
+    override fun setLocation(place: String, context: Context, map: GoogleMap) {
         val addresses :List<Address> = Geocoder(context)
             .getFromLocationName(place, 1)
         if (addresses.isNotEmpty()) {
             val latitude = addresses[0].latitude
             val longitude = addresses[0].longitude
             val placeImpl = LatLng(latitude, longitude)
-            mMap.addMarker(MarkerOptions().position(placeImpl).title(place))
-            mMap.moveCamera(CameraUpdateFactory.newLatLng(placeImpl))
-            mMap.animateCamera(CameraUpdateFactory.zoomTo(7F))
+            map.addMarker(MarkerOptions().position(placeImpl).title(place))
+            map.moveCamera(CameraUpdateFactory.newLatLng(placeImpl))
+            map.animateCamera(CameraUpdateFactory.zoomTo(7F))
         }
     }
