+++ app/src/main/res/layout/activity_maps.xml
@@ -6,4 +6,5 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".data.MapsActivity" />
+    tools:context=".ui.MapsActivity"
+    />
